% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_utils.R
\name{associatedPredictions}
\alias{associatedPredictions}
\title{Extract association model predictions with original data and metadata for plotting}
\usage{
associatedPredictions(
  SE1,
  SE2,
  assay1,
  assay2,
  sampleColumn,
  interModel,
  firstPair,
  secondPair,
  adjust = TRUE,
  outputDF = FALSE
)
}
\arguments{
\item{SE1}{SummarizedExperiment object of data from measurement-type 1 that was predicted from an interaction model.}

\item{SE2}{SummarizedExperiment object for measurement-type 1}

\item{assay1}{- name of the assay to extract and use from SE1}

\item{assay2}{- name of the assay to extract and use from SE2}

\item{sampleColumn}{- A string: the name of the column with Sample names from the metadata. 
Must be the same from SE1 and SE2. This is used for aligning SE1 and SE2}

\item{interModel}{A SummarizedExperiment object contained model coefficients and generated from any associative model function (general_Associations, scRNA_Associations, scATAC_Associations, and GeneTile_Associations).}

\item{firstPair}{Measurement name from SE1 to look at. The measurement was predicted using the measurement in secondPair, and will be adjusted for all other continuous model factors, except second pair, if adjust = TRUE.}

\item{secondPair}{Measurement name from SE2 to look at. This measurement will be the one used to predict the measurment provided to firstPair.}

\item{adjust}{A boolean, default to TRUE. Determines whether to generate an column with an adjusted measurement. The measurement will be adjusted for all factors that are not the specVariable within the model.}

\item{outputDF}{A boolea, default is FALSE, as to whether to return a summarized experiment with a seperate assay for each pair, or just a list of data.frames.}
}
\value{
A SummarizedExperiment, where each data.frame is the original data used to associated, with adjusted data (if adjust = TRUE), the model predictions, and all other metadata.
}
\description{
\code{associatedPredictions} Uses original data and an interaction model to generate group-level interaction predictions for 
                 for visualizations. It will generate a data frame of model-adjusted values from SE1, values from SE2, 
                 group-level model predictions from values in SE2, and all metadata. The dataframe for each interaction pair
                 will be saved within one assay of an SummarizedExperiment and exported.
}
\keyword{model_results}
