% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_functions.R
\name{ChromVAR_Associations}
\alias{ChromVAR_Associations}
\title{\code{ChromVAR_Associations} Model ChromVAR Z-scores as a function of features counts from a General Modality object}
\usage{
ChromVAR_Associations(
  chromSE,
  cellPopulation,
  generalSE,
  generalAssay,
  sampleColumn,
  modelFormula,
  motifList,
  generalList,
  initialSampling = 5,
  numCores = 2
)
}
\arguments{
\item{chromSE}{A ChAI ChromVAR object (SummarizedExperiment object) from \code{makeChromVAR}}

\item{cellPopulation}{- name of the assay (cell population) to extract and use from atacSE}

\item{generalSE}{A ChAI General modalityobject, generated by \code{makeChromVAR} or \code{importGeneralModality}}

\item{generalAssay}{Name of the assay from generalSE that you wish to use for modeling.}

\item{sampleColumn}{- A string: the name of the column with Sample names from the metadata. 
Must be the same from atacSE and generalSE. This is used for aligning atacSE and generalSE.}

\item{motifList}{a list of motifs to test}

\item{generalList}{A list of which rownames of generalAssay should be used for modeling associations.}

\item{initialSampling}{An integer. Default is 5. Represents the number of initial models to test when generating a null set (which is used when models fails). If your model fails often, increase this number or choose a better formula.}

\item{numCores}{Optional, the number of cores to use with multiprocessing. Default is 1.}

\item{modelFormulaa}{: Formula used for modeling. Must be in the form exp1 ~ exp2 + other factors + (1|RandomEffect). 
Exp1 will be the dummy variable for Motif Z-score, while exp2 will be the dummy variable for the general assay.}
}
\value{
A summarized experiment summarizing the output.
}
\description{
\code{ChromVAR_Associations} High-throughput modeling motif z-score changes as a function of a feature from a General Modality
}
\keyword{modeling_associations}
