% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_diagnostics.R
\name{drop1Analysis}
\alias{drop1Analysis}
\title{drop1Analysis}
\usage{
drop1Analysis(modelList, numCores = 1)
}
\arguments{
\item{modelList}{A list of models from any ChAI pilot modeling function (whether on a single modality or the association between two modalities). It should not be zero-inflated.}

\item{numCores}{The number of cores to paralelize over}
}
\value{
A dataframe of drop1Analysis results over all features, where rows are features. Failed models are not included
}
\description{
\code{drop1Analysis} Runs drop1 analysis across features in a modelList
}
\examples{
\dontrun{
  drop1Analysis(modelObj)
}

}
\keyword{evaluating_pilot}
