% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_utils.R, R/getRegulation.R
\name{exportGeneTileLinks}
\alias{exportGeneTileLinks}
\title{Export Gene-tile associations for visualization or browsing in IGV-viewer}
\usage{
exportGeneTileLinks(
  GeneTileAssociation,
  fileName,
  returnDF = FALSE,
  FDR_threshold = 0.1,
  filterFactors = NULL,
  backgroundThreshold = 0.1
)

exportGeneTileLinks(
  GeneTileAssociation,
  fileName,
  returnDF = FALSE,
  FDR_threshold = 0.1,
  filterFactors = NULL,
  backgroundThreshold = 0.1
)
}
\arguments{
\item{GeneTileAssociation}{A ChAI GeneTileAssociation model}

\item{fileName}{String describing the name of the file to write to disk.}

\item{returnDF}{Boolean. Determines whether to write to file or return a data.frame}

\item{FDR_threshold}{A number between 0 and 1, used as a significance threshold for associations between genes and tiles.}

\item{filterFactors}{A optional list of strings that describe various factors we want to remove as sources of noise. These factors should ideally represent a strong technical artifact of somekind (like sequencing depth or other). I}

\item{backgroundThreshold}{A number between 0 and 1 that is used as the FDR threshold to determine whether a controlFactor is significant, and thus a given pairwise association should be removed.}

\item{GeneTileObj}{SummarizedExperiment objectoutput of GeneTile_Associations.}

\item{filename}{The name of the file to be exported (should end in .bedpe if you want to view in IGV), if returnDF = FALSE}
}
\value{
a dataframe or nothing (writes to file)
}
\description{
\code{exportGeneTileLinks} Takes the output of GeneTile_Associations and exports a file in the bedpe format for easy browsing of gene-tile links.

\code{exportGeneTileLinksn} takes GeneTile Model and exports the significant connections to a .bedpe format or returns a data.frame.
}
\keyword{networks}
