% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object_Utils.R
\name{flattenChAI}
\alias{flattenChAI}
\title{Flatten a ChAI object from multiple matrices into one large matrix}
\usage{
flattenChAI(ChAIObject, cellPopulations = "All", metadataT = TRUE)
}
\arguments{
\item{ChAIObject}{the output of makePseudobulkRNA/normalizePseudobulk or makeChromVAR}

\item{cellPopulations}{Cell populations to keep through the combining process.}

\item{metadataT}{A flag for whether or not additional metadata is present. Should be set to FALSE. This should be removed in future versions.}
}
\value{
A SummarizedExperiment with one assay named 'counts' which contains all Sample and Cell type gene expression data.
             It also moves all the summarizedData into the colData slot of the SummarizedExperiment.
}
\description{
\code{flattenChAI} Takes a ChAI Object (ChromVAR or scRNA) and combines the matrices across cell types, while pulling in associated metadata in summarizedData and/or detectionRates
}
\keyword{data_object_manipulation}
