% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_functions.R
\name{General_Associations}
\alias{General_Associations}
\title{\code{General_Associations} Model features counts of one modality as a function of features counts from another modality (Generalized)}
\usage{
General_Associations(
  SE1,
  SE2,
  assay1,
  assay2,
  sampleColumn,
  modelFormula,
  ziFormula = ~0,
  sig1,
  sig2,
  family = stats::gaussian(),
  initialSampling = 5,
  numCores
)
}
\arguments{
\item{SE1}{SummarizedExperiment object for measurement-type 1 from \code{importGeneralModality}}

\item{SE2}{SummarizedExperiment object for measurement-type 2 from  \code{makeChromVAR} or \code{importGeneralModality}}

\item{assay1}{- name of the assay to extract and use from SE1}

\item{assay2}{- name of the assay to extract and use from SE2}

\item{sampleColumn}{- A string: the name of the column with Sample names from the metadata. 
Must be the same from SE1 and SE2. This is used for aligning SE1 and SE2}

\item{modelFormula}{: Formula used for modeling. Must be in the form exp1 ~ exp2 + other factors + (1|RandomEffect)}

\item{ziFormula}{list of  to tests from atacSE to test.}

\item{sig1}{A list of rows from SE1 to test.}

\item{sig2}{A list of rows from SE2 to test.}

\item{family}{distribution family parameter, passed to glmmTMB to describe the data's distribution.
Default is normal (gaussian()). See  \link[glmmTMB]{glmmTMB}.}

\item{initialSampling}{An integer. Default is 5. Represents the number of initial models to test when generating a null set (which is used when models fails). If your model fails often, increase this number or choose a better formula.}

\item{numCores}{Optional, the number of cores to use with multiprocessing. Default is 1.}
}
\value{
A summarized experiment summarizing the output.
}
\description{
\code{General_Associations} High-throughput zero-inlfated GLMM-based modeling for associating features between two ChAI General data objects.
}
\keyword{modeling_associations}
