% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATCHA_functions.R
\name{genestopeaks.matcha}
\alias{genestopeaks.matcha}
\title{Multi/tandem-omic workflow - Map a gene program of interest to co-accessible peaks}
\usage{
genestopeaks.matcha(
  rna.atac.combo.list.in,
  program.name.in,
  genes.in,
  save.folder.in,
  seqlevelsStyle.in = "UCSC"
)
}
\arguments{
\item{rna.atac.combo.list.in}{List of all needed datasets, formatted as described in Github example notebook}

\item{program.name.in}{Name of desired gene program}

\item{genes.in}{Genes comprising the program}

\item{save.folder.in}{Folder for saving outputs}

\item{seqlevelsStyle.in}{Style of genomic chromosome names}
}
\description{
Multi/tandem-omic workflow - Map a gene program of interest to co-accessible peaks
}
