% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_functions.R
\name{GeneTile_Associations}
\alias{GeneTile_Associations}
\title{\code{GeneTile_Associations} Model Tile accessibility as a function of gene expression}
\usage{
GeneTile_Associations(
  atacSE,
  rnaSE,
  cellPopulation,
  sampleColumn,
  continuousFormula,
  ziFormula,
  zi_threshold = 0,
  geneList,
  tileList = NULL,
  distance = 10^6,
  allCombinations = TRUE,
  initialSampling = 5,
  numCores = 2
)
}
\arguments{
\item{atacSE}{A MOCHA Tile-by-Sample Object (SummarizedExperiment)generated from getSampleTileMatrix within \code{\link[MOCHA]{MOCHA}}.}

\item{rnaSE}{A ChAI scRNA object (SummarizedExperiment object) generated by \code{makePseudobulkRNA}}

\item{cellPopulation}{- name of the assay (cell population) to extract and use from rnaSE and atacSE}

\item{sampleColumn}{- A string: the name of the column with Sample names from the metadata. 
Must be the same from atacSE and rnaSE. This is used for aligning atacSE and rnaSE}

\item{continuousFormula}{: Formula used for modeling. Must be in the form exp1 ~ exp2 + other factors + (1|RandomEffect). 
Exp1 will be the dummy variable for accessibility, while exp2 will be the dummy variable for gene expression.}

\item{ziFormula}{list of  to tests from atacSE to test.}

\item{zi_threshold}{Zero-inflated threshold ( range = 0-1), representing the fraction of samples with zeros. At or above this threshold, the zero-inflated modeling kicks in.
We don't recommend changing this unless you have fully tested out the model and understand what it is doing.}

\item{geneList}{a list of genes from rnaSE to test.}

\item{tileList}{optional list of tiles to test. Default is NULL, which will test all tiles within the provided distance of the TSS.}

\item{distance}{A number limiting the distance between tiles and genes TSS at which point associations will not be tested.}

\item{allCombinations}{A boolean, whether to test all possible pair-wise combinations of geneList and tileList. Default is TRUE. If False, it will require geneList and tileList to be the same length, and only test index-matched pairs (i.e. tileList[2] vs geneList[2] )}

\item{initialSampling}{An integer. Default is 5. Represents the number of initial models to test when generating a null set (which is used when models fails). If your model fails often, increase this number or choose a better formula.}

\item{numCores}{Optional, the number of cores to use with multiprocessing. Default is 1.}
}
\value{
A summarized experiment summarizing the output.
}
\description{
\code{GeneTile_Associations} High-throughput zero-inflated GLMM-based modeling tile accessibility as a function of gene expression
}
\keyword{modeling_associations}
