% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_diagnostics.R
\name{getConvergenceRate}
\alias{getConvergenceRate}
\title{getConvergenceRate}
\usage{
getConvergenceRate(modelList)
}
\arguments{
\item{modelList}{a list of models, output from pilotZIGLMM or pilotLMEM}
}
\value{
nothing. Prints the convergence rate out as a message.
}
\description{
\code{getConvergenceRate} Look at the model output list for any modeling function (whether unimodal or associative),
                     And get the rate at which the glmmTMB successfully converges on coefficients.
}
\examples{
\dontrun{
  getConvergenceRate(modelList)
}

}
\keyword{evaluating_pilot}
