% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object_Utils.R
\name{getDetectionRates}
\alias{getDetectionRates}
\title{Get the gene detection rates from pseudobulked scRNA}
\usage{
getDetectionRates(ChAIObj, CellType)
}
\arguments{
\item{ChAIObj}{A ChAI SummarizedExperiment, whether it's a ChAI data object or a ChAI model object}

\item{CellType}{A string describing which population-level metadata you want. The default is NULL, which will return a list of all possible types of population metadata.}
}
\value{
A data frame with the detection rate for all genes across all samples
}
\description{
\code{getDetectionRates} Get sample-wise gene detection rates (i.e. percent of cells with transcript across samples) for a given cell type from a ChAI scRNA object.
}
\examples{
\dontrun{
  getDetectionRates(dataObj, 'CD4 Naive')
}

}
\keyword{data_object_manipulation}
