% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{getEstimates}
\alias{getEstimates}
\title{get significant estimates for a given variable from any Model or Association Model}
\usage{
getEstimates(
  modelSE,
  factor = "exp2",
  FDR_threshold = 0.1,
  filterFactors = NULL,
  backgroundThreshold = 0.1,
  numCores = 1
)
}
\arguments{
\item{modelSE}{A SummarizedExperiment of model associations}

\item{factor}{A string. Default is exp2, which will pull out associations. If you used a factor in both the Zero-inflated and non-zero inflated models, you must pull each one out individualy.}

\item{FDR_threshold}{Threshold of determining significance.Default is 0.1}

\item{filterFactors}{A optional list of strings, describing the factors you do not to select for. If a given measurement is below the backgroundThreshold for that factor, it will be removed.}

\item{backgroundThreshold}{An optional background threshold that determines how stringently you want to remove measurements if they related to any of the filterFactors, based on their FDR. Standard threshold is 0.1}

\item{numCores}{If there are multiple celltypes, then it will parallelize the pulling out of estimates across cell types.}
}
\value{
data frame of coefficient values for a given factor, filtered by an FDR threshold and by the lack of relationship to other factors, if provided.
}
\description{
\code{getEstimates} Pull out a data.frame of filtered estimates with associated STD deviation, p-values, and 
             FDR for a single factor from the output SummarizedExperiment object of any modeling function.
}
\examples{
\dontrun{
  age_df <- getEstimates(model_scATAC_output, factor = 'Chr1:500-999', 
                               FDR_threshold = 0.1, backgroundThresold = 0.2)
}

}
\keyword{model_results}
