% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{getModelFactors}
\alias{getModelFactors}
\title{Get a list of factors in a model}
\usage{
getModelFactors(modelObj)
}
\arguments{
\item{modelObj}{A model object output from any ChAI modeling function (whether on a single modality or the association between two modalities)}
}
\value{
A list of strings.
}
\description{
\code{getModelFactors} Takes a model object and returns all the names of the factors used.
}
\examples{
\dontrun{
  getModelFactors(modelObj)
}

}
\keyword{model_results}
