% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{getModelTypes}
\alias{getModelTypes}
\title{Get a list of all support types of ChAI models}
\usage{
getModelTypes(optional = NULL)
}
\arguments{
\item{optional}{string}
}
\value{
A string
}
\description{
\code{getModelTypes} Returns a list of all potential model types, or checks what type a given model is.
}
\examples{
\dontrun{
  age_df <- getModelTypes()
}

}
\keyword{model_results}
