% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{getModelValues}
\alias{getModelValues}
\title{Get coefficient estimates for all factors, given one feature name}
\usage{
getModelValues(modelSE, rowName)
}
\arguments{
\item{modelSE}{A SummarizedExperiment object generated from any modeling function (model_general, model_scRNA, model_scATAC)}

\item{rowName}{A string, describing the row you want to analyze (tile from scATAC, gene for scRNA, or other from a model_general)}
}
\value{
A data.frame coefficient info by factor.
}
\description{
\code{getModelValues} Pull out a data.frame of all model values for a particular row (gene, tile, etc..).
}
\examples{
\dontrun{
  age_df <- getModelValues(runZIGLMM_output, 'Chr1:500-999')
}

}
\keyword{model_results}
