% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_diagnostics.R
\name{getPilotCoefficients}
\alias{getPilotCoefficients}
\title{getPilotCoefficients}
\usage{
getPilotCoefficients(
  pilotModelList,
  returnSummary = TRUE,
  paramInterest = "p_value"
)
}
\arguments{
\item{pilotModelList}{A list of models, output from any pilot modeling functions.}

\item{returnSummary}{A boolean, to determine whether to return the individual summaries of each model, or
to compile those results into a single data.frame for browsing. Default is TRUE (returns individual summaries)}

\item{paramInterest}{A string, only used when returnSummary = FALSE. The string describes which 
coefficient parameter you want summarized into a data.frame. 
Options include 'estimate', 'std_error', 'z_value', and 'p_value'}
}
\value{
modelList a list of outputs from glmmTMB
}
\description{
\code{getPilotCoefficients} Attempts to pull coefficients from a list of models from pilotZIGLMM or pilotLMEM. 
   Returns a list of either the coefficients for each
   model, or the error generated when attempting to get coefficients.
}
\keyword{evaluating_pilot}
