% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object_Utils.R
\name{getPopulationMetaData}
\alias{getPopulationMetaData}
\title{Get Population-level Metadata}
\usage{
getPopulationMetaData(ChAIObj, Measurement)
}
\arguments{
\item{ChAIObj}{A ChAI SummarizedExperiment, whether it's a ChAI data object or a ChAI model object}

\item{Measurement}{A string describing which population-level metadata you want. The default is NULL, which will return a list of all possible types of population metadata.}
}
\value{
A data frame with population-level information across samples and cell types
}
\description{
\code{getPopulationMetaData} Extract population-level metadata (like cell counts) from a ChAI (or MOCHA) object.
}
\examples{
\dontrun{
  getAdditionalData(modelObj, )
}

}
\keyword{data_object_manipulation}
