% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegulation.R
\name{getRegulation}
\alias{getRegulation}
\title{Combine multiple Association Models into a network of interaction features.}
\usage{
getRegulation(
  AssociationModelList,
  modelAnnotations = NULL,
  interactionDirection = NULL,
  FDR_threshold = 0.1,
  annotation_threshold = NULL,
  controlFactors = "ZI_FragNumber",
  backgroundThreshold = 0.1,
  cytoscapeOutput = FALSE
)
}
\arguments{
\item{AssociationModelList}{A list of ChAI association models. Must include at least 3 models.}

\item{modelAnnotations}{An optional list of strings describing each association model. Each string would be in the form Modality1_Modality2, and match the order of the association used. This is useful if you ran multiple general modalities. For example, modelAnnotations could be list("ATAC_RNA", "RNA_Protein", "RNA_ChromVAR", "ATAC_Protein", "ATAC_ChromVAR") for an AssociationModelList of list(GeneTile_Association, RNA_Association, RNA_Association, scATAC_Association, scATAC_Association). The order of the modalities must reflect the order of the association model - i.e. scATAC is always modeled using other modalities as a variable, and scRNA is always modeled using other modalities as a variable in the function (except for scATAC). If no modelAnnotations is given, it will simply number the association models 1,2,3,4, etc...}

\item{interactionDirection}{An optional list of the interaction direction, whicb can be useful if you want to export tables for import into cytoscape or other network visualization softwares. Using this list, it will label one modality as a source and another as a target. For example, we expect a tile's accessibility to potentially regulate a gene's expression, a transcription factor to regulate a tile accessibility, and an olink protein's concentration to influence a transcription factor's activity. So for an AssociationModelList with the model annotations of list('ATAC_RNA', 'ATAC_ChromVAR', 'ChromVAR_Olink'), we would use the interactionDirection list of list('Source_Target', 'Target_Source', 'Target_Source') to describe the direction of the regulation we expect.}

\item{FDR_threshold}{A number between 0 and 1, used as a significance threshold for associations. Any pairwise associations between modalities that have an FDR below this threshold are labeled as significant and used to build the network.}

\item{annotation_threshold}{Default is NULL, in which case it will be set to the length(AssociationModelList)-1, which will require 
All the modalities to be significantly associated with each other (i.e. A TF is sig associated to a tile and a gene, and those tiles and genes are also significantly associated with each other). If you want a less stringent threshold, you can set it to less than that manually.}

\item{controlFactors}{A optional list of strings that describe various factors we want to remove as sources of noise. These factors should ideally represent a strong technical artifact of somekind (like sequencing depth or other). If a given association model for two features is significant, it will be removed from downstream analysis. For example, if there is a significant interaction between IFNG protein levels and IRF1 ChromVAR activity, that link should be kept in the network. But if IFNG protein level is also strongly associated with limit of detection for across batches, then removing that link might make sense. You don't know if it's a significant interaction or a technical noise. This decision should be made after looking at each individual model association as well. For associations using scATAC, ZI_FragNumber should be used because drop-out will impact the degree of zero-inflation, so any significant Zero-inflated interaction term (like ZI_Age), but not continuous term, may be driven by technical noise and not the variable of interest, like Age.}

\item{backgroundThreshold}{A number between 0 and 1 that is used as the FDR threshold to determine whether a controlFactor is significant, and thus a given pairwise association should be removed.}

\item{cytoscapeOutput}{A boolean. Default is FALSE, in which case it will merge significant links together into one data.frame such that all features that pass the annotation_threshold are included. If set to TRUE, it will return a list of two data.frames, which are node and edge tables that can be imported into Cytoscape or other network visualization softwares.}
}
\value{
a dataframe or dataframes of potential networks
}
\description{
\code{getRegulation} takes multiple associative models and filters down to links
               That show a consistent relationship that is significant across associations. 
               In other words, for a given TF to bind a tile and influence gene expression, the motif should be significantly
                 associated with gene expression and tile accessibility, and the gene expression should be significantly related to tile accessibility.\
}
\keyword{networks}
