% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_utils.R
\name{getSpecificPair}
\alias{getSpecificPair}
\title{Get a data.frame with of model coefficients}
\usage{
getSpecificPair(associationPredictions, firstPair, secondPair)
}
\arguments{
\item{associationPredictions}{SummarizedExperiment object output from associatedPredictions.}

\item{firstPair}{A single measurement name from SE1}

\item{secondPair}{A single measurement name from SE2}
}
\value{
A data.frame of original data from SE1, adjusted data (if adjust = TRUE), model predictions, and all other metadata.
}
\description{
\code{getSpecificPair} Takes the output of associatedPredictions and returns a data.frame for a single combination of objects.
}
\keyword{model_results}
