% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{getVarDecomp}
\alias{getVarDecomp}
\title{Get variance decomposition from a model with only random effects}
\usage{
getVarDecomp(modelSE)
}
\arguments{
\item{modelSE}{A model object output from any ChAI modeling function (whether on a single modality or the association between two modalities)}
}
\value{
A SummarizedExperiment object.
}
\description{
\code{getVarDecomp} Takes a model object with only random effects and returns variance decomposition results
}
\examples{
\dontrun{
  getVarDecomp(modelSE)
}

}
\keyword{model_results}
