% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{importGeneralModality}
\alias{importGeneralModality}
\title{Create a ChAI General Modality object from any feature count matrix and sample metadata.}
\usage{
importGeneralModality(DataMatrix, MetaData, sampleColumn = "Sample")
}
\arguments{
\item{DataMatrix}{A data.frame, data.table, or matrix, where rows are measurements and columns are samples. Column names should match the names found within the 'Sample' column of the MetaData you provide. 
Alternatively, DataMatrix can also be named list of matrices, where all matrices are the same size, with the same row names and column names. These could be, for example, a list of gene expression across samples for multiple sorted cell types.}

\item{MetaData}{Associated metadata for each sample within the DataMatrix. Must contain a column named 'Sample' that has all the columns names within the DataMatrix}

\item{sampleColumn}{String. Name of the column that has Sample-level metadata. Default is Sample. Will overwrite any column that has Sample already present.}
}
\value{
A SummarizedExperiment with all your data inerwoven with the MetaData
}
\description{
\code{importGeneralModality} Accepts normalized data from any general modality, associated metadata, and generates a ChAI-compatible SummarizedExperiment object.
}
\examples{
\dontrun{
  olinkMat <- read.csv('olink_covid_olink.csv',row.names =1)
  olinkMeta <-  read.csv('olink_covid_olink_metadata.csv')

  #Align the column names of the data frame to the rownames of the meta data frame. 
  olinkMeta$Sample <- paste('X', olinkMeta$patientID_Visit,sep ='')
  olinkSE <- importGeneralModality(olinkMat, olinkMeta, 'Sample')
}

}
\keyword{data_import}
