% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeChromVAR.R
\name{makeChromVAR}
\alias{makeChromVAR}
\title{Generate a ChAI ChromVAR Object from MOCHA's Tile-by-Sample Matrix Object}
\usage{
makeChromVAR(
  atacSE,
  motifName = "Motifs",
  cellPopulation = NULL,
  withinCellType = FALSE,
  exportRaw = FALSE,
  numCores = 1,
  verbose = TRUE
)
}
\arguments{
\item{atacSE}{A MOCHA Tile-by-Sample Object (SummarizedExperiment)generated from getSampleTileMatrix within \code{\link[MOCHA]{MOCHA}}.}

\item{motifName}{Name of metadata slot that has motif information for analysis.}

\item{cellPopulation}{Names of cell types to analyze. Must match assay names in the SummarizedExperiment object. 
Alternative, if you want to run ChromVAR across all celltypes, you can provide the output of combineSampleTileMatrix, and set this parameter to 'counts'.}

\item{withinCellType}{Boolean. Default is FALSE, in which case accessibility across celltypes will be merged into one matrix for ChromVAR, rather than running ChromVAR seperately on the matrix for each cell type when withinCellType = TRUE.}

\item{exportRaw}{Boolean. Default is FALSE, and will export a ChAI object with ChromVAR deviations reformated for modeling. If set to true, it will either export the original chromVARDeviations object for analysis, either as a list by cell type if withinCellType is TRUE or as a single object if withinCellType is TRUE. This chromVARDeviations object can be later reformatted for ChAI using reformatChromVAR.}

\item{numCores}{Default is 1. Uses chromVAR's standard parallelization, which has memory leak errors at times. Use at your own risk.}

\item{verbose}{Boolean.}
}
\value{
A named list of ChromVAR objects.
}
\description{
\code{makeChromVAR} Runs ChromVAR on MOCHA's Tile-by-Sample Matrix object. Slow funciton.
}
\keyword{data_import}
