% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePseudobulkRNA.R
\name{makePseudobulkRNA}
\alias{makePseudobulkRNA}
\title{Generate a ChAI scRNA object (Pseudobulk scRNA)}
\usage{
makePseudobulkRNA(
  SO,
  cellTypeColumn,
  sampleColumn = "Sample",
  cellPopulations = "All",
  numCores = 2,
  Seurat_format = "SYMBOL",
  TxDb = NULL,
  OrgDb = NULL,
  normalize = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{SO}{Seurat Object}

\item{cellTypeColumn}{The column of the Seurat object with cell type information}

\item{sampleColumn}{The column of the Seurat object with sample information}

\item{cellPopulations}{A list of cell type names to extract or the word 'All' (which is default). Determines which cell populations you want to aggregate and analyze from the Seurat object.}

\item{numCores}{The number of cores to parallelize this operation over.}

\item{Seurat_format}{A string, describing the gene name format of the Seurat object. This is used to annotate gene loci, and convert IDs. See documentation for AnnotationDbi::mapIds to identify formats. Default is 'SYMBOL', but 'ENSEMBL' is also common.}

\item{TxDb}{A Transcript database to be used for identifying gene locations. Must be installed already. Both TxDb and OrgDb must be provided to annote the gene locations.}

\item{OrgDb}{An Organism database to match up gene names and locations. Must be installed already. Both TxDb and OrgDb must be provided to annote the gene locations.}

\item{normalize}{A boolean to determine whether or not to return DESeq2-normalized counts, or raw aggregated counts. Default is TRUE.}

\item{verbose}{Boolean. Default is TRUE, and will print messages.}
}
\value{
A SummarizedExperiment carrying pseudobulked average expression per 1000 cells for each cell type.
}
\description{
\code{makePseuduoulkRNA} pseodubulks a Seurat object by sample and cell type into a SummarizedExperiment, similar to MOCHA
}
\keyword{data_import}
