% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object_Utils.R
\name{mergeChAI}
\alias{mergeChAI}
\title{Merge ChAI objects of similar data types into one.}
\usage{
mergeChAI(ChAIObjectList)
}
\arguments{
\item{ChAIObjectList}{A list of ChAI SummarizedExperiments, each one with the same history (i.e. all are makePseudobulkRNA, normalizePseudobulk, makeChromVAR, or importGeneralModality.}
}
\value{
A ChAI object with all the samples combined (including any summarizedData and detectionRates, if necessary)
}
\description{
\code{mergeChAI} Takes the ChAI objects that are the same type (scRNA, ChromVAR, or general) and merges them. Allows for running data import in batches before merging.
}
\keyword{data_object_manipulation}
