% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{model_ChromVAR}
\alias{model_ChromVAR}
\title{High-throughput Modeling of ChromVAR Motif scores}
\usage{
model_ChromVAR(
  chromSE,
  cellPopulation,
  modelFormula,
  initialSampling = 5,
  verbose = FALSE,
  numCores = 2
)
}
\arguments{
\item{chromSE}{A SummarizedExperiment object generated from chromVAR via ChAI::makeChromVAR}

\item{cellPopulation}{The name of the assay to model within the SummarizedExperiment.}

\item{modelFormula}{The formula to use, in the
format (exp ~ factors). All factors must be found in column names
of the chromSE metadata. modelFormula must start with 'exp' as the response.
See \link[glmmTMB]{glmmTMB}.}

\item{initialSampling}{Size of data to use for pilot}

\item{verbose}{Set TRUE to display additional messages. Default is FALSE.}

\item{numCores}{integer. Number of cores to parallelize across.}
}
\value{
results a SummarizedExperiment containing LMEM results. Assays are metrics related to the model coefficients,
         including the Estimate, Std_Error, df, t_value, p_value. Within each assay, each row corresponds to each row of
         the SummarizedExperiment and columns correspond to each fixed effect variable within the model.
         Any row metadata from the chromSE (see rowData(chromSE)) is preserved in the output. 
         The Residual matrix and the variance of the random effects are saved in the metadata slot of the output.
}
\description{
\code{model_ChromVAR} Runs linear mixed-effects modeling for
  chromVAR Z-scores or deviations, derived from a MOCHA object using makeChromVAR, and modeled using \code{\link[glmmTMB]{glmmTMB}}
}
\examples{
\dontrun{
  modelList <- model_ChromVAR(chromSE,
   cellPopulation = "CD16 Mono",
    modelFormula = NULL,
    initialSampling = 5,
    verbose = FALSE,
    numCores = 1
 )
}

}
\keyword{modeling_individual}
