% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{model_General}
\alias{model_General}
\title{High-through modeling (Generalized)}
\usage{
model_General(
  ExperimentObj,
  assayName = "General",
  modelFormula,
  ziFormula = ~0,
  family = stats::gaussian(),
  initialSampling = 5,
  verbose = FALSE,
  numCores = 2
)
}
\arguments{
\item{ExperimentObj}{A SummarizedExperiment object generated from chromVAR, or other.
It is expected to contain normally distributed data without zero-inflation.}

\item{assayName}{The name of the assay to model within the SummarizedExperiment.}

\item{modelFormula}{The formula to use, in the
format (exp ~ factors). All factors must be found in column names
of the ExperimentObj metadata. modelFormula must start with 'exp' as the response.
See \link[glmmTMB]{glmmTMB}.}

\item{ziFormula}{A formula for zero inflation. By default, zero-inflated modeling is turned off by setting ziFormula = ~ 0}

\item{family}{distribution family parameter, passed to glmmTMB to describe the data's distribution.
Default is normal (gaussian()). See  \link[glmmTMB]{glmmTMB}.}

\item{initialSampling}{Size of data to use for pilot}

\item{verbose}{Set TRUE to display additional messages. Default is FALSE.}

\item{numCores}{integer. Number of cores to parallelize across.}
}
\value{
results a SummarizedExperiment containing modeling results. Assays are metrics related to the model coefficients,
         including the Estimate, Std_Error, df, t_value, p_value. Within each assay, each row corresponds to each row of
         the SummarizedExperiment and columns correspond to each fixed effect variable within the model.
         Any row metadata from the ExperimentObject (see rowData(ExperimentObj)) is preserved in the output. 
         The Residual matrix and the variance of the random effects are saved in the metadata slot of the output.
         If multiple cell types are provided, then a named list of the above objects will be returned, one for each object.
}
\description{
\code{model_General} Runs generalized GLM modeling for
  continuous, non-zero inflated data using \code{\link[glmmTMB]{glmmTMB}}
}
\examples{
\dontrun{
  modelList <- model_General(ExperimentObj,
   assayName = "z",
    modelFormula = NULL,
    initialSampling = 5,
    verbose = FALSE,
    numCores = 1
 )
}

}
\keyword{modeling_individual}
