% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{model_scATAC}
\alias{model_scATAC}
\title{High-throughut Modeling of scATAC data}
\usage{
model_scATAC(
  atacSE,
  cellPopulation,
  modelFormula = NULL,
  ziFormula = ~0,
  zi_threshold = 0,
  initialSampling = 5,
  verbose = FALSE,
  numCores = 2
)
}
\arguments{
\item{atacSE}{A SummarizedExperiment object generated from
getSampleTileMatrix.}

\item{cellPopulation}{Name of a cell type.}

\item{modelFormula}{The formula for the continuous data that should be used within glmmTMB. It should be in the
format (exp ~ factors). All factors must be found in column names
of the atacSE metadata, except for CellType, FragNumber and CellCount, which will be extracted from the atacSE.
modelFormula must start with 'exp' as the response.
See \link[glmmTMB]{glmmTMB}.}

\item{ziFormula}{The formula for the zero-inflated data that should be used within glmmTMB. It should be in the
format ( ~ factors). All factors must be found in column names
of the atacSE colData metadata, except for CellType, FragNumber and CellCount, which will be extracted from the atacSE.}

\item{zi_threshold}{Zero-inflated threshold ( range = 0-1), representing the fraction of samples with zeros. At or above this threshold, the zero-inflated modeling kicks in.}

\item{initialSampling}{Size of data to use for pilot}

\item{verbose}{Set TRUE to display additional messages. Default is FALSE.}

\item{numCores}{integer. Number of cores to parallelize across.}
}
\value{
results a SummarizedExperiment containing ZI-GLMM results
}
\description{
\code{model_scATAC} High-throughput modeling of tile accessibility from scATAC \code{\link[glmmTMB]{glmmTMB}}.
}
\examples{
\dontrun{
  modelList <- model_scATAC(STM[c(1:1000),], 
                 cellPopulation = 'CD16 Mono',
                 modelFormula = exp~ Age + Sex + days_since_symptoms + (1|PTID), 
                 ziFormula = ~ 0 + FragNumber + Age, 
                 verbose = TRUE, 
                 numCores = 35 )
}

}
\keyword{modeling_individual}
