% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{model_scRNA}
\alias{model_scRNA}
\title{High-throughout Modeling of Gene Expression}
\usage{
model_scRNA(
  rnaSE,
  cellPopulation = NULL,
  modelFormula = NULL,
  ziFormula = ~0,
  zi_threshold = 0,
  family = "negativeBinomial2",
  detectionThreshold = 0.01,
  expressionThreshold = 0,
  cellCountThreshold = 10,
  initialSampling = 5,
  verbose = FALSE,
  numCores = 2
)
}
\arguments{
\item{rnaSE}{A SummarizedExperiment object generated from normalizePseudobulk}

\item{cellPopulation}{Name(s) of a cell type to be modeled, or 'all' if you want to model across all cell types.}

\item{modelFormula}{The formula to use, in the
format (exp ~ factors). All factors must be found in column names
of the ExperimentObj metadata. modelFormula must start with 'exp' as the response.
See \link[glmmTMB]{glmmTMB}.}

\item{ziFormula}{A formula for zero inflation. By default, zero-inflated modeling is turned off by setting ziFormula = ~ 0}

\item{family}{String. Can be 'negativeBinomial1', 'negativeBinomial2', or 'poisson'. Default is "negativeBinomial2".}

\item{detectionThreshold}{A number between 0 and 1, representing the mean detection rate threshold for a given gene to be modeled. 
This detection rate is calculated for each sample and cell type during makePseudobulkRNA, and represents the percentage of cells that have a transcript for a given gene. Over all samples, the average has to be above this to be modeled. Default is 0.01.}

\item{expressionThreshold}{A number greater than zero, representing the expression threshold for modeling. A given gene, on average across all samples, expressed above this threshold. The default is 0.}

\item{cellCountThreshold}{The minimum number of cells in a given pseudobulk for the pseudobulk to be included in analysis. If fewer than this number of cells are found, then the sample will be dicarded The number of cells within the pseudobulked scRNA. Default is 10 cells.}

\item{initialSampling}{Size of data to use for pilot}

\item{verbose}{Set TRUE to display additional messages. Default is FALSE.}

\item{numCores}{integer. Number of cores to parallelize across.}
}
\value{
results a SummarizedExperiment containing modeling results. Assays are metrics related to the model coefficients,
         including the Estimate, Std_Error, df, t_value, p_value. Within each assay, each row corresponds to each row of
         the SummarizedExperiment and columns correspond to each fixed effect variable within the model.
         Any row metadata from the ExperimentObject (see rowData(ExperimentObj)) is preserved in the output. 
         The Residual matrix and the variance of the random effects are saved in the metadata slot of the output.
         If multiple cell types are provided, then a named list of the above objects will be returned, one for each object.
}
\description{
\code{model_scRNA} Runs high-throughout GLM modeling of gene expression using \code{\link[glmmTMB]{glmmTMB}}
}
\examples{
\dontrun{
  modelList <- model_scRNA(rnaSE,
   cellPopulation = "CD14_Mono"
    modelFormula = "exp ~ treatment + (1|patientID)",
    initialSampling = 5,
    verbose = FALSE,
    numCores = 1
 )
}

}
\keyword{modeling_individual}
