% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{modelPredictions}
\alias{modelPredictions}
\title{Get model predictions, original measurements, and metadata for a given feature}
\usage{
modelPredictions(
  dataObject,
  assay1,
  measurement,
  specVariable,
  modelObject,
  adjust = TRUE,
  ignoreVariable = NULL
)
}
\arguments{
\item{dataObject}{The SummarizedExperiment object with real data.}

\item{assay1}{Name of the assay to analyze}

\item{measurement}{The specific measurement (gene, motif, tile, protein, etc...) that you want to look at within dataObject}

\item{specVariable}{The variable(s) of interest. If other fixed effects are present, they will be adjusted for. Must be actual model factors. Use getModelFactors to see the names of model factors in a given model object.}

\item{modelObject}{a SummarizedOutput object from modelPredict}

\item{adjust}{A boolean, default to TRUE. Determines whether to generate an column with an adjusted measurement. The measurement will be adjusted for all factors that are not the specVariable within the model.}

\item{ignoreVariable}{Default is NULL. use this to specific whether there are any variables that you do not want to adjust for.}
}
\value{
A summarized experiment where each assay is the data.frame of metadata, original measurement, predicted measurement, and adjusted measurement (option) for each measurement provided.
}
\description{
\code{modelPredictions} Uses the raw data object and the model prediction from a ChAI modeling function to generate a data.frame of raw and predicted features values for plotting.
}
\keyword{model_results}
