% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATCHA_functions.R
\name{multiprogram.crossdataset.TFprioritize.matcha}
\alias{multiprogram.crossdataset.TFprioritize.matcha}
\title{Across gene programs, prioritize consensus transcription factors that drive multiple programs across all datasets}
\usage{
multiprogram.crossdataset.TFprioritize.matcha(
  folder.in,
  gene.list.in,
  n.tfs.per.program = 10,
  min.tf.degree = 2,
  graph.layout = "fr",
  plot.width = 8,
  plot.height = 8,
  plot.font.size = 12
)
}
\arguments{
\item{folder.in}{Folder in which outputs of all datasets are saved}

\item{gene.list.in}{List object containing all gene programs, formatted as in example notebook}

\item{n.tfs.per.program}{Top and bottom n transcription factors to retain for each program}

\item{min.tf.degree}{Minimum number of programs to which a transcription factor has to be linked in order to be retained}

\item{graph.layout}{Layout algorithm for plotting transcription factor - gene program network. Can be any layout compatible with ggraph}

\item{plot.width}{Desired width of output plot}

\item{plot.height}{Desired height of output plot}

\item{plot.font.size}{Desired font size of output plot}
}
\description{
Across gene programs, prioritize consensus transcription factors that drive multiple programs across all datasets
}
