% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATCHA_functions.R
\name{peakstoTFs.matcha}
\alias{peakstoTFs.matcha}
\title{Multi/tandem-omic workflow - Quantify accessibility of transcription factor motifs within program-linked co-accessible peaks}
\usage{
peakstoTFs.matcha(rna.atac.combo.list.in, save.folder.in, program.name.in)
}
\arguments{
\item{rna.atac.combo.list.in}{List of all needed datasets, formatted as described in Github example notebook}

\item{save.folder.in}{Folder in which the output of genestopeaks.matcha was saved}

\item{program.name.in}{Name of desired gene program}
}
\value{
Signac object with a new chromVAR assay corresponding to transcription factor motif accessibility within program-linked co-accessible peaks
}
\description{
Multi/tandem-omic workflow - Quantify accessibility of transcription factor motifs within program-linked co-accessible peaks
}
