% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pilots.R
\name{pilot_chromVAR}
\alias{pilot_chromVAR}
\title{Test formulas for modeling ChromVAR motif Z-scores}
\usage{
pilot_chromVAR(
  chromSE,
  cellPopulation = NULL,
  modelFormula,
  numCores = 1,
  featureSample = 100,
  verbose = FALSE
)
}
\arguments{
\item{chromSE}{A SummarizedExperiment-type object generated from makeChromVAR.}

\item{cellPopulation}{Name of a cell type within the ChAI ChromVAR Object}

\item{modelFormula}{The formula to use with glmmTMB, in the
format (exp ~ factors). All factors must be found in column names
of the ExperimentObj metadata.}

\item{numCores}{integer. Number of cores to parallelize across.}

\item{featureSample}{integer or vector of characters describing the features to test. 
If integer, stratified sampling will occur to identify features to test, with a minimum of 10. 
If a character vector, then all strings must align to features you want to test.}

\item{verbose}{Set TRUE to display additional messages. Default is FALSE.}
}
\value{
modelList a list of outputs from glmmTMB
}
\description{
\code{pilot_chromVAR} Function for testing out GLM formulas on the ChromVAR Motif Z-scores generated via ChAI and MOCHA
  Runs a given formula on a subset of the data, and returns the model results. This is meant to help during the model selection process. \code{\link[glmmTMB]{glmmTMB}}.
}
\keyword{pilot_modeling}
