% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pilots.R
\name{pilot_General}
\alias{pilot_General}
\title{Test formulas for modeling a general modality}
\usage{
pilot_General(
  ExperimentObj,
  assayName = NULL,
  modelFormula,
  ziFormula = ~0,
  zi_threshold = 0,
  numCores = 1,
  family = stats::gaussian(),
  featureSample = 100,
  verbose = FALSE
)
}
\arguments{
\item{ExperimentObj}{A SummarizedExperiment-type object generated from importGeneralModality}

\item{assayName}{a character string, matching the name of an assay within the SummarizedExperiment. 
The assay named will be used for modeling.}

\item{modelFormula}{The formula to use with glmmTMB, in the
format (exp ~ factors). All factors must be found in column names
of the ExperimentObj metadata.}

\item{ziFormula}{A formula for zero inflation. By default, zero-inflated modeling is turned off by setting ziFormula = ~ 0}

\item{zi_threshold}{A threshold above which the ziFormula is used. If zero-inflation is below this value (e.g. 0.1, or 1/10 values are zero), then those values are censored. This threshold is ignored if ziFormula = ~ 0}

\item{numCores}{integer. Number of cores to parallelize across.}

\item{family}{distribution family parameter, passed to glmmTMB to describe the data's distribution.
Default is normal (gaussian()). See  \link[glmmTMB]{glmmTMB}.}

\item{featureSample}{integer or vector of characters describing the features to test. 
If integer, stratified sampling will occur to identify features to test, with a minimum of 10. 
If a character vector, then all strings must align to features you want to test.}

\item{verbose}{Set TRUE to display additional messages. Default is FALSE.}
}
\value{
modelList a list of outputs from glmmTMB
}
\description{
\code{pilot_General} Generalized Function for testing out zero-inflated and standard GLM formula on a general modality imported via \code{importGeneralModality}
  Runs a given formula on a subset of the data, and returns the model results. This is meant to help during the model selection process. \code{\link[glmmTMB]{glmmTMB}}.
}
\keyword{pilot_modeling}
