% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_pilots.R
\name{pilot_General_Associations}
\alias{pilot_General_Associations}
\title{Test formulas for modeling one modality's feature counts as a function of another modality's feature counts (Generalized)}
\usage{
pilot_General_Associations(
  SE1,
  SE2,
  assay1,
  assay2,
  sampleColumn,
  modelFormula,
  ziFormula = ~0,
  sig1,
  sig2,
  family = stats::gaussian(),
  numCores = 2,
  samplingNumber = 5
)
}
\arguments{
\item{SE1}{SummarizedExperiment object for measurement-type 1 from \code{importGeneralModality}}

\item{SE2}{SummarizedExperiment object for measurement-type 2 from  \code{makeChromVAR} or \code{importGeneralModality}}

\item{assay1}{- name of the assay to extract and use from SE1}

\item{assay2}{- name of the assay to extract and use from SE2}

\item{sampleColumn}{- A string: the name of the column with Sample names from the metadata. 
Must be the same from SE1 and SE2. This is used for aligning SE1 and SE2}

\item{modelFormula}{Formula used for modeling. Must be in the form exp1 ~ exp2 + other factors + (1|RandomEffect)}

\item{ziFormula}{A formula for zero inflation. By default, zero-inflated modeling is turned off by setting ziFormula = ~ 0}

\item{sig1}{A list of rows from SE1 to test.}

\item{sig2}{A list of rows from SE2 to test.}

\item{family}{distribution family parameter, passed to glmmTMB to describe the data's distribution.
Default is normal (gaussian()). See  \link[glmmTMB]{glmmTMB}.}

\item{numCores}{Number of cores for glmmTMB to parallelize over.}

\item{samplingNumber}{An integer or a list of pairs (e.g. Protein1_TranscriptionFactor1). If an integer, it will randomly pull that many pairs from all possible combinations of sig1 and sig2, returning the model outputs. 
If a list of pairs is provided as a list of strings, it will only test those. These pairs should be a subset of all possible pairs of sig1 and sig2.}
}
\value{
A named list of model outputs
}
\description{
\code{pilot_General_Associations} Function for testing formulas for zero-inflated or standard GLM modeling of
                  one modality's feature counts as a function of another modality's feature counts. 
                  Runs a given formula on a subset of the data, and returns the model results. 
                  This is meant to help during the model selection process. 
                  The second general modality should be roughly normally distributed. \code{\link[glmmTMB]{glmmTMB}}.
}
\keyword{pilot_associations}
