% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_pilots.R
\name{pilot_GeneTile}
\alias{pilot_GeneTile}
\title{Test formulas for modeling tile accessibility as a function of gene expression.}
\usage{
pilot_GeneTile(
  atacSE,
  rnaSE,
  cellPopulation,
  sampleColumn,
  continuousFormula,
  ziFormula,
  zi_threshold = 0,
  geneList,
  tileList = NULL,
  distance = 10^6,
  samplingNumber = 5,
  numCores = 2
)
}
\arguments{
\item{atacSE}{A MOCHA Tile-by-Sample Object (SummarizedExperiment)generated from getSampleTileMatrix within \code{\link[MOCHA]{MOCHA}}.}

\item{rnaSE}{A ChAI scRNA object (SummarizedExperiment object) generated by \code{makePseudobulkRNA}}

\item{cellPopulation}{- name of the assay (cell population) to extract and use from rnaSE and atacSE}

\item{sampleColumn}{- A string: the name of the column with Sample names from the metadata. 
Must be the same from atacSE and rnaSE. This is used for aligning atacSE and rnaSE}

\item{continuousFormula}{: Formula used for modeling. Must be in the form exp1 ~ exp2 + other factors + (1|RandomEffect). 
Exp1 will be the dummy veriable for accessibility, while exp2 will be the dummy variable for gene expression.}

\item{ziFormula}{list of  to tests from atacSE to test.}

\item{zi_threshold}{Zero-inflated threshold ( range = 0-1), representing the fraction of samples with zeros. At or above this threshold, the zero-inflated modeling kicks in.
We don't recommend changing this unless you have fully tested out the model and understand what it is doing.}

\item{geneList}{a list of genes from rnaSE to test.}

\item{tileList}{optional list of tiles to test. Default is NULL, which will test all tiles within the provided distance of the TSS.}

\item{distance}{A number limiting the distance between tiles and genes TSS at which point associations will not be tested.}

\item{samplingNumber}{An integer or a list of pairs (e.g. Protein1_TranscriptionFactor1). 
 If an integer, it will randomly pull that many pairs from all possible combinations of sig1 and sig2, returning the model outputs. 
If a list of pairs is provided as a list of strings, it will only test those. These pairs should be a subset of all possible pairs of sig1 and sig2.}

\item{numCores}{Optional, the number of cores to use with multiprocessing. Default is 2.}
}
\value{
A named list of model outputs
}
\description{
\code{pilot_GeneTile} Test formulas for zero-inflated GLMM-based modeling of tile accessibility 
                   as a function of gene expression. \code{\link[glmmTMB]{glmmTMB}}.
}
\examples{
\dontrun{
  modelList <- pilot_TilesGeneLinks(STM, 
                           'CD16 Mono',
                           exp~ Age + Sex + days_since_symptoms + (1|PTID),
                            ~ FragNumber, verbose = TRUE )
}

}
\keyword{pilot_associations}
