% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pilots.R
\name{pilot_scATAC}
\alias{pilot_scATAC}
\title{Test formulas for modeling tile accessibility (from MOCHA)}
\usage{
pilot_scATAC(
  atacSE,
  cellPopulation = NULL,
  modelFormula = NULL,
  ziFormula = ~0,
  zi_threshold = 0,
  verbose = FALSE,
  numCores = 1,
  featureSample = 10
)
}
\arguments{
\item{atacSE}{A MOCHA Tile-by-Sample Object (SummarizedExperiment)generated from getSampleTileMatrix within \code{\link[MOCHA]{MOCHA}}.}

\item{cellPopulation}{Name of a cell type within the atacSE}

\item{modelFormula}{The formula for the continuous data that should be used within glmmTMB. It should be in the
format (exp ~ factors). All factors must be found in column names
of the atacSE metadata, except for CellType, FragNumber and CellCount, which will be extracted from the atacSE.
modelFormula must start with 'exp' as the response.
See \link[glmmTMB]{glmmTMB}.}

\item{ziFormula}{The formula for the zero-inflated data that should be used within glmmTMB. It should be in the
format ( ~ factors). All factors must be found in column names
of the atacSE colData metadata, except for CellType, FragNumber and CellCount, which will be extracted from the atacSE.
FragNumber and CellCounts will be log10 normalized within the function.}

\item{zi_threshold}{Zero-inflated threshold ( range = 0-1), representing the fraction of samples with zeros. At or above this threshold, the zero-inflated modeling kicks in.}

\item{verbose}{Set TRUE to display additional messages. Default is FALSE.}

\item{numCores}{Number of cores to provide to glmmTMB for parallelization.}

\item{featureSample}{integer or vector of characters describing the features to test. If integer, stratified sampling will occur to identify features to test, with a minimum of 10. If a character vector, then all strings must align to features you want to test.}
}
\value{
results model results
}
\description{
\code{pilot_scATAC} Functions for testing out zero-inflated GLMM formulas on the data from MOCHA. Runs a given formula on a subset of the data, and returns the model results. This is meant to help during the model selection process. \code{\link[glmmTMB]{glmmTMB}}.
}
\examples{
\dontrun{
  modelList <- pilot_scATAC(STM, 
                           'CD16 Mono',
                           exp~ Age + Sex + days_since_symptoms + (1|PTID),
                            ~ FragNumber, verbose = TRUE )
}

}
\keyword{pilot_modeling}
