% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_pilots.R
\name{pilot_scATAC_Associations}
\alias{pilot_scATAC_Associations}
\title{Test formulas for modeling tile accessibility as a function of another modality's feature counts.}
\usage{
pilot_scATAC_Associations(
  atacSE,
  cellPopulation,
  generalSE,
  generalAssay,
  sampleColumn,
  continuousFormula,
  ziFormula = ~0 + FragNumber,
  zi_threshold = 0,
  tileList,
  generalList,
  samplingNumber = 5,
  numCores = 2
)
}
\arguments{
\item{atacSE}{A MOCHA Tile-by-Sample Object (SummarizedExperiment)generated from getSampleTileMatrix within \code{\link[MOCHA]{MOCHA}}.}

\item{cellPopulation}{Name of a cell type within the TSAM_Object}

\item{generalSE}{A ChAI General modalityobject, generated by \code{makeChromVAR} or \code{importGeneralModality}}

\item{generalAssay}{Name of the assay within generalSE that should be used.}

\item{sampleColumn}{- A string: the name of the column with Sample names from the metadata. 
Must be the same from generalSE and atacSE. This is used for aligning generalSE and atacSE.}

\item{continuousFormula}{The formula for the continuous data that should be used within glmmTMB. It should be in the
format (exp ~ factors). All factors must be found in column names
of the TSAM_Object metadata, except for CellType, FragNumber and CellCount, which will be extracted from the atacSE.
modelFormula must start with 'exp' as the response.
See \link[glmmTMB]{glmmTMB}.}

\item{ziFormula}{The formula for the zero-inflated data that should be used within glmmTMB. It should be in the
format ( ~ factors). All factors must be found in column names
of the atacSE colData metadata, except for CellType, FragNumber and CellCount, which will be extracted from the TSAM_Object.
FragNumber and CellCounts will be log10 normalized within the function.}

\item{zi_threshold}{Zero-inflated threshold ( range = 0-1), representing the fraction of samples with zeros. At or above this threshold, the zero-inflated modeling kicks in.}

\item{tileList}{A list of tiles, as strings in the format 'ChrX:1000-1499' to use. All tiles in this list will be tested for association with all measurements within generalList.}

\item{generalList}{A list of measurements from generalSE to associate with all tiles in the tileList.}

\item{samplingNumber}{An integer or a list of pairs (e.g. Protein1_TranscriptionFactor1). If an integer, it will randomly pull that many pairs from all possible combinations of sig1 and sig2, returning the model outputs. 
If a list of pairs is provided as a list of strings, it will only test those. These pairs should be a subset of all possible pairs of sig1 and sig2.}

\item{numCores}{Optional, the number of cores to use with multiprocessing. Default is 2.}
}
\value{
A named list of model outputs
}
\description{
\code{pilot_scATAC_Associations} Function for testing formulas for zero-inflated modeling of
                  tile accessibility as a function of features from a General Modality.
                  Runs a given formula on a subset of the data, and returns the model results. 
                  This is meant to help during the model selection process. 
                  The General modality should be roughly normally distributed. \code{\link[glmmTMB]{glmmTMB}}.
}
\examples{
\dontrun{
}

}
\keyword{pilot_associations}
