% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pilots.R
\name{pilot_scRNA}
\alias{pilot_scRNA}
\title{Test formulas for modeling gene expression}
\usage{
pilot_scRNA(
  rnaSE,
  cellPopulation = NULL,
  modelFormula = NULL,
  ziFormula = ~0,
  zi_threshold = 0,
  verbose = FALSE,
  family = "negativeBinomial2",
  detectionThreshold = 0.1,
  expressionThreshold = 0,
  cellCountThreshold = 0,
  numCores = 1,
  featureSample = 100
)
}
\arguments{
\item{rnaSE}{A SummarizedExperiment object generated from makePseudobulkRNA.}

\item{cellPopulation}{Name of a cell type within the rnaSE}

\item{modelFormula}{The formula for the continuous data that should be used within glmmTMB. It should be in the
format (exp ~ factors). All factors must be found in column names
of the rnaSE metadata, except for CellType, FragNumber and CellCount, which will be extracted from the rnaSE.
modelFormula must start with 'exp' as the response.
See \link[glmmTMB]{glmmTMB}.}

\item{verbose}{Set TRUE to display additional messages. Default is FALSE.}

\item{family}{String. Can be 'negativeBinomial1', 'negativeBinomial2', or 'poisson'. Default is "negativeBinomial2".}

\item{detectionThreshold}{A number between 0 and 1, representing the mean detection rate threshold for a given gene to be modeled. 
This detection rate is calculated for each sample and cell type during makePseudobulkRNA, and represents the percentage of cells that have a transcript for a given gene. Over all samples, the average has to be above this to be modeled. Default is 0.01.}

\item{expressionThreshold}{A number greater than zero, representing the expression threshold for modeling. A given gene, on average across all samples, expressed above this threshold. The default is 0.}

\item{cellCountThreshold}{The minimum number of cells in a given pseudobulk for the pseudobulk to be included in analysis. If fewer than this number of cells are found, then the sample will be dicarded The number of cells within the pseudobulked scRNA. Default is 10 cells.}

\item{numCores}{Number of cores to provide to glmmTMB for parallelization.}

\item{featureSample}{integer or vector of characters describing the features to test. If integer, stratified sampling will occur to identify features to test, with a minimum of 10. If a character vector, then all strings must align to features you want to test.}
}
\value{
results model results
}
\description{
\code{pilot_scRNA} Function for testing out GLM formulas on the pseudobulked RNA via ChAI.
  Runs a given formula on a subset of the data, and returns the model results. This is meant to help during the model selection process. \code{\link[glmmTMB]{glmmTMB}}.
}
\examples{
\dontrun{
  modelList <- pilot_scRNA(rnaSE, 
                           'CD16 Mono',
                           exp~ Age + Sex + days_since_symptoms + (1|PTID),
                             verbose = TRUE )
}

}
\keyword{pilot_modeling}
