% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_pilots.R
\name{pilot_scRNA_Associations}
\alias{pilot_scRNA_Associations}
\title{Test formulas for modeling gene expression as a funciton of another modality's feature counts.}
\usage{
pilot_scRNA_Associations(
  rnaSE,
  cellPopulation,
  generalSE,
  generalAssay,
  sampleColumn,
  modelFormula,
  ziFormula = ~0,
  cellCountThreshold = 10,
  family = "negativeBinomial2",
  geneList,
  generalList = NULL,
  samplingNumber = 100,
  numCores = 10
)
}
\arguments{
\item{rnaSE}{A ChAI scRNA object (SummarizedExperiment object) generated by \code{makePseudobulkRNA}}

\item{cellPopulation}{- name of the assay (cell population) to extract and use from rnaSE}

\item{generalSE}{A ChAI General modalityobject, generated by \code{makeChromVAR} or \code{importGeneralModality}}

\item{generalAssay}{Name of the assay from generalSE that you wish to use for modeling.}

\item{sampleColumn}{- A string: the name of the column with Sample names from the metadata. 
Must be the same from generalSE and rnaSE. This is used for aligning generalSE and rnaSE}

\item{modelFormula}{: Formula used for modeling. Must be in the form exp1 ~ exp2 + other factors + (1|RandomEffect), where 
exp1 is the dummy variable for gene expression and exp2 is the dummy variable for the rownames from generalSE.}

\item{ziFormula}{A formula for zero inflation. By default, zero-inflated modeling is turned off by setting ziFormula = ~ 0}

\item{cellCountThreshold}{The minimum number of cells in a given pseudobulk for the pseudobulk to be included in analysis. If fewer than this number of cells are found, then the sample will be dicarded The number of cells within the pseudobulked scRNA. Default is 10 cells.}

\item{family}{String. Can be 'negativeBinomial1', 'negativeBinomial2', or 'poisson'. Default is "negativeBinomial2".}

\item{geneList}{a list of genes from rnaSE to test.}

\item{generalList}{A list of which rownames of generalAssay should be used for modeling associations.}

\item{samplingNumber}{An integer or a list of pairs (e.g. Protein1_TranscriptionFactor1). If an integer, it will randomly pull that many pairs from all possible combinations of sig1 and sig2, returning the model outputs. 
If a list of pairs is provided as a list of strings, it will only test those. These pairs should be a subset of all possible pairs of sig1 and sig2.}

\item{numCores}{Optional, the number of cores to use with multiprocessing. Default is 2.}
}
\value{
A named list of model outputs
}
\description{
\code{pilot_scRNA_Associations} Function for testing formulas for zero-inflated modeling of gene expression as a function of features from a General Modality. 
             Runs a given formula on a subset of the data, and returns the model results. 
               This is meant to help during the model selection process. 
                The General modality should be roughly normally distributed. \code{\link[glmmTMB]{glmmTMB}}.
}
\examples{
\dontrun{
  modelList <- pilot_scRNA(scRNA_Object, 
                           'CD16 Mono',
                           exp~ Age + Sex + days_since_symptoms + (1|PTID),
                             verbose = TRUE )
}

}
\keyword{pilot_associations}
