% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_diagnostics.R
\name{pilotPredictions}
\alias{pilotPredictions}
\title{pilotPredictions}
\usage{
pilotPredictions(modelList, specVariable = NULL, returnList = FALSE)
}
\arguments{
\item{modelList}{a list of models, output from pilotZIGLMM or pilotLMEM}

\item{specVariable}{Default is NULL. If used, provide a String that matches one of the variables in your model. The experimental measurement (gene, tile, or other) will be adjusted for all other variables except for this one.}

\item{returnList}{Boolean. Default is FALSE. Determines if you want a list of data.frames for each feature that was modeled, or one combined data.frame of values across features/models}
}
\value{
list of dataframes containing metadata, and model predictions (if successful) or one combined data frame across features.
}
\description{
\code{pilotPredictions} extract model predictions, metadata, and adjusted measurements from the output of pilot_scRNA, pilot_scATAC, or pilot_General
}
\examples{
\dontrun{
  predictedDF <- plotZIModels(modelList,returnMatrix = TRUE)
}

}
\keyword{evaluating_pilot}
