% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_diagnostics.R
\name{plotPilotFits}
\alias{plotPilotFits}
\title{plotPilotFits}
\usage{
plotPilotFits(modelList, fileName)
}
\arguments{
\item{modelList}{A list of models from any ChAI pilot modeling function (whether on a single modality or the association between two modalities)}

\item{fileName}{Name of the pdf to export. The function will add .pdf to the end of this string.}
}
\value{
NULL Save a pdf under the given name with residuals plotted.
}
\description{
\code{plotPilotFits} Runs DHARMa::simulateResiduals for zero-inflated models.
}
\examples{
\dontrun{
  pilot_performance(modelObj)
}

}
\keyword{evaluating_pilot}
