% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeChromVAR.R
\name{reformatChromVAR}
\alias{reformatChromVAR}
\title{Reformats raw output from makeChromVAR}
\usage{
reformatChromVAR(chromVARList, selectDev = FALSE)
}
\arguments{
\item{chromVARList}{The output of makeChromVAR, either as a cell type list (withinCellType and exportList are TRUE, which produces a a named list of ChromVAR objects for each cell type, or a single ChromVAR object run across multiple cell types when withinCellType is FALSE}

\item{selectDev}{A boolean, with default to FALSE, that decides whether or not deviations or Z score values will be used in the output object. Default of FALSE chooses Z-scores.}
}
\value{
A SummarizedExperiment of either z-scores or deviations across all cell types, (each assay is a cell type), similar to the format of a MOCHA object for scATAC data, or the output of makePseudobulkRNA. 
             This format is necessary for using ChromVAR with modeling and association functions (model_General, scATAC_Associations, scRNA_Associations, general_Associations)
}
\description{
\code{reformatChromVAR} Takes a celltype list of chromVARDeviations or a single chromVARDeviations object over all cell types and reformats it into a ChAI object where each assay is a cell type
}
\keyword{data_import}
