% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object_Utils.R
\name{renameCellTypes}
\alias{renameCellTypes}
\title{Rename cell types (or assays) within a ChAI Object}
\usage{
renameCellTypes(Object, oldNames, newNames)
}
\arguments{
\item{Object}{A SummarizedExperiment object from MOCHA or ChAI}

\item{oldNames}{A list of cell type names that you want to change.}

\item{newNames}{A list of new cell type names to replace the old names with.}
}
\value{
A MOCHA SampleTile object with new cell types.
}
\description{
\code{renameCellTypes} Allows you to modify the cell type names for a MOCHA SampleTileObject, ChAI pseudobulk scRNA object, or just any general assay from
                   the assay names, GRanges column names, and summarizedData (within the metadata), all at once.
}
\examples{
\dontrun{
  chromSE <- renameCellTypes(chromSE, 'CD16_Mono', 'CD16 Mono')
}

}
\keyword{data_object_manipulation}
