% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_functions.R
\name{scRNA_Associations}
\alias{scRNA_Associations}
\title{\code{scRNA_Associations} Model gene expression as a function of features counts from a General Modality object}
\usage{
scRNA_Associations(
  rnaSE,
  cellPopulation,
  generalSE,
  generalAssay,
  sampleColumn,
  modelFormula,
  ziFormula = ~0,
  cellCountThreshold = 10,
  family = "negativeBinomial2",
  geneList,
  generalList = NULL,
  initialSampling = 5,
  numCores = 2
)
}
\arguments{
\item{rnaSE}{A ChAI scRNA object (SummarizedExperiment object) generated by \code{makePseudobulkRNA}}

\item{cellPopulation}{- name of the assay (cell population) to extract and use from rnaSE and atacSE}

\item{generalSE}{A ChAI General modalityobject, generated by \code{makeChromVAR} or \code{importGeneralModality}}

\item{generalAssay}{Name of the assay from generalSE that you wish to use for modeling.}

\item{sampleColumn}{- A string: the name of the column with Sample names from the metadata. 
Must be the same from atacSE and rnaSE. This is used for aligning atacSE and rnaSE}

\item{modelFormula}{: Formula used for modeling. Must be in the form exp1 ~ exp2 + other factors + (1|RandomEffect), where 
exp1 is the dummy variable for gene expression and exp2 is the dummy variable for the rownames from generalSE.}

\item{ziFormula}{list of  to tests from atacSE to test.}

\item{cellCountThreshold}{The minimum number of cells in a given pseudobulk for the pseudobulk to be included in analysis. If fewer than this number of cells are found, then the sample will be dicarded The number of cells within the pseudobulked scRNA. Default is 10 cells.}

\item{family}{String. Can be 'negativeBinomial1', 'negativeBinomial2', or 'poisson'. Default is "negativeBinomial2".}

\item{geneList}{a list of genes from rnaSE to test.}

\item{generalList}{A list of which rownames of generalAssay should be used for modeling associations.}

\item{initialSampling}{An integer. Default is 5. Represents the number of initial models to test when generating a null set (which is used when models fails). If your model fails often, increase this number or choose a better formula.}

\item{numCores}{Optional, the number of cores to use with multiprocessing. Default is 2.}
}
\value{
A summarized experiment summarizing the output.
}
\description{
\code{scRNA_Associations} High-throughput GLMM-based modeling to associate gene expression to another modality, while controlling for technical confounders
}
\keyword{modeling_associations}
