% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATCHA_functions.R
\name{singleprogram.crossdataset.TFprioritize.matcha}
\alias{singleprogram.crossdataset.TFprioritize.matcha}
\title{For a single gene program, prioritize consensus transcription factors that drive the program across all datasets}
\usage{
singleprogram.crossdataset.TFprioritize.matcha(
  folder.in,
  name.in,
  plot.width = 6,
  plot.height = 6,
  plot.font.size = 12
)
}
\arguments{
\item{folder.in}{Folder in which outputs of all datasets are saved}

\item{name.in}{Name of gene program}

\item{plot.width}{Desired width of output plot}

\item{plot.height}{Desired height of output plot}

\item{plot.font.size}{Desired font size of output plot}
}
\description{
For a single gene program, prioritize consensus transcription factors that drive the program across all datasets
}
