% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object_Utils.R
\name{subsetChAI}
\alias{subsetChAI}
\title{Subset a ChAI object}
\usage{
subsetChAI(ChAIObject, subsetBy, groupList)
}
\arguments{
\item{ChAIObject}{the output of makePseudobulkRNA or normalizePseudobulk}

\item{subsetBy}{A string that is either 'CellType', 'Assay', or the name of column that has sample names in the Seurat object's metadata}

\item{groupList}{The list of cell types or samples that you want to keep.}
}
\value{
A SummarizedExperiment with only the cell types or samples of interest.
}
\description{
\code{subsetChAI} Takes the output of makePseudobulkRNA, normalizedPseudobulk, makeChromVAR, or importGeneralModality
}
\examples{
\dontrun{
  chromSE2 <- subsetChAI(subChrom, subsetBy = 'COVID_status', 'Negative')
}


}
\keyword{data_object_manipulation}
