% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_diagnostics.R
\name{summarizeResFits}
\alias{summarizeResFits}
\title{summarizeResFits}
\usage{
summarizeResFits(
  summaryOfFits,
  threshold = 0.05,
  uncorrected = TRUE,
  correctionType = "fdr",
  verboseOutput = TRUE,
  na.rm = FALSE
)
}
\arguments{
\item{summaryOfFits}{A data frame, generated by testPilotFits}

\item{threshold}{A p-value threshold used for the various tests, to check whether the model for a feature failed a test. The standard is p = 0.05.}

\item{uncorrected}{A boolean, indicating whether to use the nominal p-value for each statistical tests, or to convert those p-values to a corrected p-value before calculating the percentage of models that failed a given test. Default is TRUE. If FALSE, then corrected p-values are used.}

\item{correctionType}{A string, documenting which type of p-value adjustment to use, if uncorrected = FALSE. See R documentation on p.adjust for options.}

\item{verboseOutput}{A boolean flag, that determines whether to print the results as a percentage, or return a named vector with numeric values.}

\item{na.rm}{Some tests may generate NAs. This removes those NAs when calculating the percent failed. Default is FALSE.}
}
\value{
A named lists, showing what percentage of models failed a given test.
}
\description{
\code{summarizeResFits} Summarized the output of testPilotFits, allowing a quick summary of how many models (as a percentage) failed which tests.
}
\examples{
\dontrun{
  summarizeResFits(model_fit_data_frame)
}

}
\keyword{evaluating_pilot}
