% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_diagnostics.R
\name{testPilotFits}
\alias{testPilotFits}
\title{testPilotFits}
\usage{
testPilotFits(modelList, numCores = 1)
}
\arguments{
\item{modelList}{A list of models from any ChAI pilot modeling function (whether on a single modality or the association between two modalities)}

\item{numCores}{Number of cores to parallelize over.}
}
\value{
NULL. Save a pdf under the given name with residuals plotted.
}
\description{
\code{testPilotFits } Runs DHARMa::simulateResiduals for zero-inflated models, and runs a variety of tests on each model to evaluate model fit.
}
\examples{
\dontrun{
  pilot_performance(modelObj)
}

}
\keyword{evaluating_pilot}
