% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATCHA_functions.R
\name{TFstoRankedTFs.atac.rna.matcha}
\alias{TFstoRankedTFs.atac.rna.matcha}
\title{Multi/tandem-omic workflow - Identify transcription factors whose expression and motif accessibility within program-linked co-accessible peaks associates with program expression}
\usage{
TFstoRankedTFs.atac.rna.matcha(
  rna.atac.combo.list.in,
  save.folder.in,
  program.name.in,
  genes.in
)
}
\arguments{
\item{rna.atac.combo.list.in}{List of all needed datasets, formatted as described in Github example notebook}

\item{save.folder.in}{Folder in which the output of genestopeaks.matcha was saved}

\item{program.name.in}{Name of desired gene program}

\item{genes.in}{Genes comprising the gene program}
}
\value{
Dataframe of prioritized transcription factors
}
\description{
Multi/tandem-omic workflow - Identify transcription factors whose expression and motif accessibility within program-linked co-accessible peaks associates with program expression
}
