% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATCHA_functions.R
\name{TFstoRankedTFs.rna.only.matcha}
\alias{TFstoRankedTFs.rna.only.matcha}
\title{Transcriptomic-only workflow - Identify transcription factors whose expression associates with program expression}
\usage{
TFstoRankedTFs.rna.only.matcha(
  rna.only.list.in,
  save.folder.in,
  program.name.in,
  genes.in
)
}
\arguments{
\item{rna.only.list.in}{List of all needed datasets, formatted as described in Github example notebook}

\item{save.folder.in}{Folder in which the output of genestopeaks.matcha was saved}

\item{program.name.in}{Name of desired gene program}

\item{genes.in}{Genes comprising the gene program}
}
\value{
Dataframe of prioritized transcription factors
}
\description{
Transcriptomic-only workflow - Identify transcription factors whose expression associates with program expression
}
