% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object_Utils.R
\name{transformChAI}
\alias{transformChAI}
\title{Run a mathematical transform on a ChAI object.}
\usage{
transformChAI(ChAIObject, transformType = "clr", ...)
}
\arguments{
\item{ChAIObject}{A SummarizedExperiment generated from ChAI (makeChromVAR, normalizePseudobulk, or importGeneralModality)}

\item{transformType}{A string that must be either 'log10p' for a log10+1 transform, 'logp' for log e + 1, 
'CLR' for a center-log ratio transform, and 'meanShift' (data - rowMeans(data)). 
 Additional Cox-Box transforms inclde r, rv, r2, rv2, l, lv, l2, and lv2, based on this Nature paper. 
 https://www.nature.com/articles/s41598-019-41315-w & the corresponding package 'countTransformers'}
}
\value{
A ChAI object with the data transformed
}
\description{
\code{transformChAI} For associations, abnormally distributed data can break models. transformChAI does several specific transformations to data in ChAI objects, 
                 so that they can be used in associative modeling functions.
}
\examples{
\dontrun{
  tObject <- transformChAI(geneExpression, transformType = 'CLR')
}

}
\keyword{data_object_manipulation}
