% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_utils.R
\name{visualizeAssociations}
\alias{visualizeAssociations}
\title{\code{visualizeAssociations}}
\usage{
visualizeAssociations(
  interAct,
  dataType1 = TRUE,
  dataTypeName = "TF",
  threshold = 20,
  max.overlaps = 10,
  verbose = FALSE,
  rastr = FALSE
)
}
\arguments{
\item{interAct}{A dataframe of coefficients related to a modality, extracted from multiModalModeling or other. This should be prefiltered for significance (FDR < 0.1)}

\item{dataType1}{Boolean flag. This determines whether it should look at the Modality1 or Modality2 column (i.e. the first or second assay given to multiModalIntegration)}

\item{dataTypeName}{A string to label the modality in the plots generated.}

\item{threshold}{A number, describing how many points to label. By default, it will label the top 20 points.}

\item{max.overlaps}{The maximum number of overlapping labels, a parameter which is passed to ggrepel when labeling points on the plot.}

\item{verbose}{Boolean flag to determine verbosity.}

\item{rastr}{A boolean flag to determine whether to rasterize the pdfs. Requires ggrastr to be installed.}
}
\value{
A ggplot object, showing number of interactions on the y-axis and the rank order of measurements on the x-axis 
   (Measurements rank-ordered by the number of interactions they have with the other modality.)
}
\description{
\code{visualizeAssociations} Takes the output of multiModalModeling (or other) and visualizes the top interaction features
 for one of the modalities.
}
\keyword{model_results}
