% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_utils.R
\name{volcanoPairs}
\alias{volcanoPairs}
\title{\code{volcanoPairs}}
\usage{
volcanoPairs(
  interModel,
  FDR_threshold = 0.05,
  topLimit = 20,
  max.overlaps = 10,
  returnDF = FALSE,
  rastr = FALSE,
  addSignificanceLine = TRUE
)
}
\arguments{
\item{interModel}{The summarizedExperiment output of any association modeling funciton (scATAC_Associations, scRNA_Associations, general_Associations, & GeneTile_Associations)}

\item{FDR_threshold}{A number, which thresholds which points are considered significant. It will also add a red line to the plot to mark significance.}

\item{topLimit}{A number, limiting how many points to label. By default, it will label the top 20 points by FDR, if there are more than 20 points.}

\item{max.overlaps}{The maximum number of overlapping labels, a parameter which is passed to ggrepel when labeling points on the plot.}

\item{returnDF}{Boolean flag to determine whether to return the ggplot or a data.frame of the information. If FALSE, it will return a ggplot and not a data.frame.}

\item{rastr}{A boolean flag to determine whether to rasterize the pdfs. Requires ggrastr to be installed.}

\item{addSignificanceLine}{Boolean flag to determine whether or not to add a line at the FDR significance threshold.}
}
\value{
A ggplot object, showing number of interactions on the y-axis and the rank order of measurements on the x-axis 
   (Measurements rank-ordered by the number of interactions they have with the other modality.)
}
\description{
\code{volcanoPairs} Takes the output of multiModalModeling (or other) and generates a volcano plot of Estimate vs -log10(FDR) for that pair
}
\keyword{model_results}
