% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object_Utils.R
\name{getAdditionalData}
\alias{getAdditionalData}
\title{Extract additional metadata from a ChAI data or model Object}
\usage{
getAdditionalData(ChAIObj, feature = NULL)
}
\arguments{
\item{ChAIObj}{A ChAI SummarizedExperiment, whether it's a ChAI data object or a ChAI model object}

\item{feature}{The type of data you want. If null, it will return the list of all options for that objct. 
Options include 'summarizedData', which contains population-specific metadata like cell counts, 
'detectionRates', which contains the percentage of a population with transcript of a given gene (scRNA only)
 'Residuals', which contains a SummarizedExperiment of Residuals after modeling
 or 'Variance'/'RandomEffectVariance' which will return a data.frame of variance from a model.}
}
\value{
A list of additional data available within a given ChAI object, or a specific data object from that ChAI object.
}
\description{
\code{getAdditionalData} Returns a wide range of metadata from ChAI objects, including
                 population-level metadata (scRNA, and scATAC), gene detection rates within a sample-celltype (scRNA),
                  residuals (models), and random effect variance (models)
}
\examples{
\dontrun{
  getAdditionalData(modelObj, )
}

}
\keyword{data_object_manipulation}
