% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_functions.R
\name{scATAC_Associations}
\alias{scATAC_Associations}
\title{Model tile accessibility as a function of features counts from a General Modality object}
\usage{
scATAC_Associations(
  atacSE,
  cellPopulation,
  generalSE,
  generalAssay,
  sampleColumn,
  continuousFormula,
  ziFormula = ~0 + FragNumber,
  zi_threshold = 0,
  tileList,
  generalList,
  initialSampling = 5,
  numCores = 2
)
}
\arguments{
\item{atacSE}{A MOCHA Tile-by-Sample Object (SummarizedExperiment)generated from getSampleTileMatrix within \code{\link[MOCHA]{MOCHA}}.}

\item{cellPopulation}{- name of the assay (cell population) to extract and use from atacSE}

\item{generalSE}{A ChAI General modalityobject, generated by \code{makeChromVAR} or \code{importGeneralModality}}

\item{generalAssay}{Name of the assay from generalSE that you wish to use for modeling.}

\item{sampleColumn}{- A string: the name of the column with Sample names from the metadata. 
Must be the same from atacSE and generalSE. This is used for aligning atacSE and generalSE.}

\item{continuousFormula}{: Formula used for modeling. Must be in the form exp1 ~ exp2 + other factors + (1|RandomEffect). 
Exp1 will be the dummy variable for accessibility, while exp2 will be the dummy variable for gene expression.}

\item{ziFormula}{list of  to tests from atacSE to test.}

\item{zi_threshold}{Zero-inflated threshold ( range = 0-1), representing the fraction of samples with zeros. At or above this threshold, the zero-inflated modeling kicks in.
We don't recommend changing this unless you have fully tested out the model and understand what it is doing.}

\item{tileList}{a list of tiles to test.}

\item{generalList}{A list of which rownames of generalAssay should be used for modeling associations.}

\item{initialSampling}{An integer. Default is 5. Represents the number of initial models to test when generating a null set (which is used when models fails). If your model fails often, increase this number or choose a better formula.}

\item{numCores}{Optional, the number of cores to use with multiprocessing. Default is 1.}
}
\value{
A summarized experiment summarizing the output.
}
\description{
\code{scATAC_Associations} High-throughput zero-inflated GLMM-based modeling tile accessibility as a function of features from a General Modality
}
\keyword{modeling_associations}
